LayoutTests/imported/w3c/web-platform-tests/content-security-policy/script-src/script-src-report-only-policy-works-with-external-hash-policy-expected.txt


FAIL Should fire securitypolicyviolation event assert_equals: expected "report" but got "enforce"
FAIL External script in a script tag with matching SRI hash should run. assert_true: External script ran. expected true got false
